// src/recipe/thirdparty/types.ts
export interface User {
  // https://github.com/supertokens/core-driver-interface/wiki#third-party-user
  id: string
  timeJoined: number
  email: string
  thirdParty: {
    id: string
    userId: string
  }
}
// ...
export interface RecipeInterface {
  getUserById(input: { userId: string; userContext: any }): Promise<User | undefined>

  getUsersByEmail(input: { email: string; userContext: any }): Promise<User[]>

  getUserByThirdPartyInfo(input: {
    thirdPartyId: string
    thirdPartyUserId: string
    userContext: any
  }): Promise<User | undefined>

  signInUp(input: {
    thirdPartyId: string
    thirdPartyUserId: string
    email: string
    userContext: any
  }): Promise<{ status: 'OK'; createdNewUser: boolean; user: User }>
}